Creating the courses page so that tecahers can upload files for students to use
There were several problems such as the upload of the course not coming up, or the pdf file not displaying, or the redirection onto antoher page. In order to fix this, the team had to ensure that the redirection was occuring from the course.html page and not teacher.html page. Then create 2 new databases that stored information on the topic, subject, file, description.